// src/api/models/MessageThreadSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageThreadSchema = {
    user_id: number;
    user_name: string;
    user_role: string;
    last_message: string;
    last_message_time: string;
    unread_count: number;
};