// src/api/models/OutMessageSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OutMessageSchema = {
    receiver_id: number;
    content: string;
    subject?: (string | null);
    parent_message_id?: (number | null);
    id: number;
    sender_id: number;
    is_read: boolean;
    created_at: string;
    updated_at: string;
};